fix(components/packages): install `ng2-dragula` if used in source code
